window-slot: remove unneeded code special casing desktop
We are special casing to open the desktop, so in case we try to
open the desktop in this window slot, and was already opened, we
were opening it again.

I cannot see a reason why we would need to open it anyway if was
already opened, and the test I did works fine.

So let's remove it for now, so we remove another special casing of the
desktop.