properties-window: use app info for "Open With" visibility
We were special casing desktop icon files to not show the Open With
tab in the properties window.

However, this is just equal to what we need to do to display Open With
in the context menus, where we check if the app has a possible
application handling it or not.

Do the same here so we can remove another desktop dependency.